refactor: local-storage-service -> database/core
localStorage is not a good name for an indexedDB database as it already
stands for a different storage that we may also use.

The "core" of our database interacting service is responsible for
directly manipulating the database. It is not responsible for
understanding and transforming the models or for interacting with the
remote server.